refactored folders structure
